fix: export types were published incorrectly (#29548)
