+++ packages/devextreme-angular/src/common/export/excel/index.ts
@@ -1,4 +1,6 @@
 export {
+    CellAddress,
+    CellRange,
     DataGridCell,
     DataGridExportOptions,
     exportDataGrid,
