+++ packages/devextreme-vue/src/common/export/excel.ts
@@ -1,4 +1,6 @@
 export {
+    CellAddress,
+    CellRange,
     DataGridCell,
     DataGridExportOptions,
     exportDataGrid,
