+++ packages/devextreme/js/core/element.d.ts
@@ -7,12 +7,14 @@ export interface ElementWrapper<T extends Element> { }
 export interface ElementsArrayWrapper<T extends Element> { }
 /**
  * @docid
+ * @hidden
  * @type HTMLElement|SVGElement|JQuery
  */
 export type DxElement<T extends Element = HTMLElement> = {} extends Condition ? T : ElementWrapper<T>;
 
 /**
  * @docid
+ * @hidden
  * @type HTMLElement|SVGElement|JQuery
  */
 export type UserDefinedElement<T extends Element = Element> = {} extends Condition ? T : ElementWrapper<T> | T;
