+++ packages/devextreme/js/core/utils/deferred.d.ts
@@ -28,6 +28,7 @@ export interface PromiseType<T> { }
 /**
  * @docid
  * @type Promise<void>
+ * @hidden
  * @namespace DevExpress.core.utils
  */
