+++ packages/devextreme/js/data/data_source.d.ts
@@ -23,6 +23,7 @@ export type Options<
 
 /**
  * @docid
+ * @hidden
  * @type Store|DataSource|DataSourceOptions|string|Array<any>|null
  */
 export type DataSourceLike<TItem, TKey = any> =
