+++ packages/devextreme/js/events/events.types.d.ts
@@ -84,7 +84,7 @@ export interface EventType { }
 /**
  * @docid
  * @type EventObject|jQuery.Event
- *
+ * @hidden
  */
 export type DxEvent<TNativeEvent = Event> = {} extends EventType
   ? (EventObject<TNativeEvent> & TNativeEvent)
